pyproject.toml
[tool.poetry]
name = "predictor ruleta online con react"
version = "0.1.0"
description = "spin predictor con react render since serverside"
authors = ["jonathan-escritorio <joni92@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
asgiref = "3.8.1"
django = "5.1"
psycopg2-binary = "^2.9.9"
sqlparse = "0.5.1"
tzdata = "2024.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
firebase-admin = "^6.6.0"
coreapi = "^2.3.3"
pyyaml = "^6.0.2"
drf-yasg = "^1.21.8"
numpy = "1.26.4"
tensorflow-io-gcs-filesystem = "0.31.0"
tensorflow ="2.17.0"
tensorflow-intel="2.17.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
pytest = "^8.3.5"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

